store: add composeEndpoints to composerApi
