+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -1,3 +1,5 @@
 export { lookupDatastreamDistro } from './dataStreamLookup';
 export { getBlueprintsPath } from './getBlueprintsPath';
 export { getCloudConfigs } from './getCloudConfigs';
+export { readComposes } from './readComposes';
+export { toComposerComposeRequest } from './toComposerComposeRequest';
