+++ src/store/api/backend/onprem/composerApi/index.ts
@@ -1,298 +1,21 @@
-import path from 'path';
-
-// Note: for the on-prem version of the frontend we have configured
-// this so that we check `node_modules` and `pkg/lib` for packages.
-// To get around this for the hosted service, we have configured
-// the `tsconfig` to stubs of the `cockpit` and `cockpit/fsinfo`
-// modules. These stubs are in the `src/test/mocks/cockpit` directory.
-// We also needed to create an alias in vitest to make this work.
-import cockpit from 'cockpit';
-import { fsinfo } from 'cockpit/fsinfo';
-
-import { IMAGE_MODE } from '@/constants';
-import {
-  ComposeBlueprintApiArg,
-  ComposeBlueprintApiResponse,
-  ComposeResponse,
-  ComposesResponseItem,
-  GetBlueprintComposesApiArg,
-  GetBlueprintComposesApiResponse,
-  GetComposesApiArg,
-  GetComposesApiResponse,
-  GetComposeStatusApiArg,
-  GetComposeStatusApiResponse,
-  OpenScapProfile,
-} from '@/store/api/backend';
-import { onPremQueryHandler } from '@/store/api/shared';
-
 import { architectureEndpoints } from './architecture';
 import { blueprintEndpoints } from './blueprints';
-import { getBlueprintsPath } from './helpers';
+import { composeEndpoints } from './composes';
 import { oscapEndpoints } from './oscap';
 import { podmanEndpoints } from './podman';
 import { workerEndpoints } from './worker';
 
 import { emptyComposerApi } from '../emptyComposerApi';
-import { assertComposeResponse, assertComposeStatus } from '../typeguards';
-import {
-  type ComposerCreateBlueprintRequest,
-  type ComposerComposeRequest as ComposeRequest,
-  type ComposerImageRequest,
-  type ComposerCustomizations as Customizations,
-  type ComposerImageTypes as ImageTypes,
-} from '../types';
-
-const readComposes = async (bpID: string) => {
-  const blueprintsDir = await getBlueprintsPath();
-  let composes: ComposesResponseItem[] = [];
-  const bpInfo = await fsinfo(
-    path.join(blueprintsDir, bpID),
-    ['entries', 'mtime'],
-    {
-      superuser: 'try',
-    },
-  );
-  const bpEntries = Object.entries(bpInfo.entries || {});
-  for (const entry of bpEntries) {
-    if (entry[0] === `${bpID}.json`) {
-      continue;
-    }
-    const composeReq = await cockpit
-      .file(path.join(blueprintsDir, bpID, entry[0]))
-      .read();
-    composes = [
-      ...composes,
-      {
-        id: entry[0],
-        request: JSON.parse(composeReq),
-        created_at: new Date(entry[1]!.mtime * 1000).toString(),
-        blueprint_id: bpID,
-      },
-    ];
-  }
-  return composes;
-};
-
-export const toComposerComposeRequest = (
-  blueprint: ComposerCreateBlueprintRequest,
-  distribution: string,
-  image_requests: ComposerImageRequest[],
-): ComposeRequest => {
-  // subscription, users & openscap are the only options
-  // that aren't compatibile with the on-prem customizations,
-  // so we have to handle those separately
-  const { subscription, openscap, users, ...hostedCustomizations } =
-    blueprint.customizations;
-
-  const customizations: Customizations = {
-    ...hostedCustomizations,
-  };
-
-  if (openscap) {
-    customizations.openscap = {
-      // the casting here is fine since compliance isn't available on-prem
-      profile_id: (openscap as OpenScapProfile).profile_id,
-    };
-  }
-
-  if (users) {
-    customizations.users = users.map((user) => {
-      const { ssh_key, groups, password } = user;
-      return {
-        name: user.name,
-        ...(groups && { groups: groups }),
-        ...(password && { password: password }),
-        ...(ssh_key && { key: ssh_key }),
-      };
-    });
-  }
-
-  if (subscription) {
-    customizations!.subscription = {
-      organization: subscription.organization.toString(),
-      activation_key: subscription['activation-key'],
-      server_url: subscription['server-url'],
-      base_url: subscription['base-url'],
-      rhc: subscription.rhc,
-      insights: subscription.insights,
-      insights_client_proxy: subscription.insights_client_proxy,
-    };
-  }
-
-  let distro: string | undefined = distribution;
-  if (distro === IMAGE_MODE) {
-    distro = undefined;
-  }
-
-  let bootc = undefined;
-  if (blueprint.bootc) {
-    bootc = blueprint.bootc;
-  }
-
-  return {
-    distribution: distro,
-    bootc: bootc,
-    customizations,
-    image_requests: image_requests.map((ir) => ({
-      architecture: ir.architecture,
-      image_type: ir.image_type as ImageTypes,
-      repositories: [],
-      upload_targets: [
-        {
-          type: ir.upload_request.type,
-          upload_options: ir.upload_request.options,
-        },
-      ],
-    })),
-  };
-};
 
 export const composerApi = emptyComposerApi.injectEndpoints({
   endpoints: (builder) => {
     return {
       ...architectureEndpoints(builder),
       ...blueprintEndpoints(builder),
+      ...composeEndpoints(builder),
       ...oscapEndpoints(builder),
       ...workerEndpoints(builder),
       ...podmanEndpoints(builder),
-      composeBlueprint: builder.mutation<
-        ComposeBlueprintApiResponse,
-        ComposeBlueprintApiArg
-      >({
-        queryFn: onPremQueryHandler(
-          async ({ queryArgs: { id: filename }, baseQuery }) => {
-            const blueprintsDir = await getBlueprintsPath();
-            const file = cockpit.file(
-              path.join(blueprintsDir, filename, `${filename}.json`),
-            );
-            const contents = await file.read();
-            const parsed = JSON.parse(contents);
-
-            const blueprint = parsed as ComposerCreateBlueprintRequest;
-            const composes: ComposeResponse[] = [];
-            for (const ir of blueprint.image_requests) {
-              if (ir.upload_request.type === 'aws.s3') {
-                // this differs to crc because the on-prem backend
-                // can actually understand a `local` image type.
-                // We can build this locally rather than sending it
-                // to an s3 bucket.
-                ir.upload_request.type = 'local';
-              }
-
-              // this request gets saved to the local storage and needs to
-              // match the hosted format
-              const crcComposeRequest = {
-                ...blueprint,
-                image_requests: [ir],
-              };
-
-              const result = await baseQuery({
-                url: '/compose',
-                method: 'POST',
-                body: JSON.stringify(
-                  // since this is the request that gets sent to the cloudapi
-                  // backend, we need to modify it slightly
-                  toComposerComposeRequest(
-                    blueprint,
-                    crcComposeRequest.distribution,
-                    [ir],
-                  ),
-                ),
-                headers: {
-                  'content-type': 'application/json',
-                },
-              });
-
-              if (result.error) {
-                throw result.error;
-              }
-
-              const { id } = assertComposeResponse(result.data);
-              await cockpit
-                .file(path.join(blueprintsDir, filename, id))
-                .replace(JSON.stringify(crcComposeRequest));
-              composes.push({ id });
-            }
-
-            return composes;
-          },
-        ),
-      }),
-      getComposes: builder.query<GetComposesApiResponse, GetComposesApiArg>({
-        queryFn: onPremQueryHandler(async () => {
-          const blueprintsDir = await getBlueprintsPath();
-          const info = await fsinfo(blueprintsDir, ['entries'], {
-            superuser: 'try',
-          });
-          let composes: ComposesResponseItem[] = [];
-          const entries = Object.entries(info.entries || {});
-          for (const entry of entries) {
-            composes = composes.concat(await readComposes(entry[0]));
-          }
-          return {
-            meta: {
-              count: composes.length,
-            },
-            links: {
-              first: composes.length > 0 ? composes[0].id : '',
-              last: composes.length > 0 ? composes[composes.length - 1].id : '',
-            },
-            data: composes,
-          };
-        }),
-      }),
-      getBlueprintComposes: builder.query<
-        GetBlueprintComposesApiResponse,
-        GetBlueprintComposesApiArg
-      >({
-        queryFn: onPremQueryHandler(async ({ queryArgs }) => {
-          const composes = await readComposes(queryArgs.id);
-          return {
-            meta: {
-              count: composes.length,
-            },
-            links: {
-              first: composes.length > 0 ? composes[0].id : '',
-              last: composes.length > 0 ? composes[composes.length - 1].id : '',
-            },
-            data: composes,
-          };
-        }),
-      }),
-      getComposeStatus: builder.query<
-        GetComposeStatusApiResponse,
-        GetComposeStatusApiArg
-      >({
-        queryFn: onPremQueryHandler(async ({ queryArgs, baseQuery }) => {
-          const result = await baseQuery({
-            url: `/composes/${queryArgs.composeId}`,
-            method: 'GET',
-          });
-
-          if (result.error) {
-            throw result.error;
-          }
-
-          const data = assertComposeStatus(result.data);
-          const blueprintsDir = await getBlueprintsPath();
-          const info = await fsinfo(blueprintsDir, ['entries'], {
-            superuser: 'try',
-          });
-
-          const entries = Object.entries(info.entries || {});
-          for (const bpEntry of entries) {
-            const request = await cockpit
-              .file(path.join(blueprintsDir, bpEntry[0], queryArgs.composeId))
-              .read();
-            return {
-              image_status: data.image_status,
-              request: JSON.parse(request),
-            };
-          }
-
-          throw new Error('Compose not found');
-        }),
-      }),
     };
   },
   // since we are inheriting some endpoints,
@@ -323,3 +46,6 @@ export const {
   usePodmanImagesQuery,
   useLazyPodmanImagesQuery,
 } = composerApi;
+
+// re-export this for testing
+export { toComposerComposeRequest } from './helpers';
